Course 2 Play IT Safe Conducting a security audit
After completing lessons and quizzes, I was tasked with conducting a security audit for a fictional business. 

The situation is that the fictional company's IT Manager wants to conduct a security audit. Her concerns: 
- lack of security and risk management 
- securing their infrastructure
- identifying and mitigating risks, threats, and vulnerabilities to critical assets
- comply with EU GDPR

The background of the fictional business is that they're a U.S. toy manufacturer and retailer with one location acting as its corporate office, a storefront, and a warehouse. The fictional company has developed quite the online presence, attracting domestic and international customers. The online growth is why the IT Manager has concerns. 

The IT Manager implemented NIST CSF, established an audit scope and goal, listed IT-managed assets, and completed a risk assessment. The goal is to list the risks and fines the company faces with its current security posture. 

This document is the scope, goals, and risk assessment report.